Create exercise-2-debugging-practice
Instructions
Look at the code in the code editor on the right. There are errors in all of the lines of code. Fix the code so that it runs without errors.

Warning: The output in your program should match the example output shown below exactly, character for character, even spaces and symbols should be identical, otherwise the tests won't pass.