Create day-4-project-rock-paper-scissors.py
Head over to replit.com
https://replit.com/@appbrewery/rock-paper-scissors-start?v=1
Coding Rooms is very handy for our exercises. However, the end-of-day projects are more free-form. You can experiment and play with this project. Customise it. Get as creative as you like. There is no exact output for you to match. There are no tests for you to pass. These assignments are not graded.